rpki-validator: convert binary rib to csv
The README states that a binary rib dump should be passed to
rpki-validator.sh.
The rpki-validator.sh converts it to csv and afterwards does the
analysis.

Uncomment the code to convert the rib dump to a csv file and add an
if-clause to only convert it a file named <passed-filename>.formatted
doesn't exist.